+++ index.html
@@ -81,6 +81,6 @@
     <div id="confetti-container"></div>
 
     <script src="https://cdn.jsdelivr.net/npm/tone@14.7.77/build/Tone.min.js"></script>
-    <script src="js/app.js"></script>
+    <script type="module" src="js/ui.js"></script>
   </body>
 </html>
